README
Making any change without consulting could end up breaking the patch, I am not responsible for any mishaps that happen once the current formula is edited.  

The current formula was voted on by the community, and was specifically made to be not intefere with the current gameplay of new vegas. In-fact it's added much more balance, the addition of cut stun grenades for example, people will be able to temporarily blind explosive griefers This patch was voted on for how great it made their experience, and I highly recommend keeping it as it is, at least trial the file in it's current state , i would be most greatful.